FIX: rename folders and update file paths for AI engine models
- Changed folder names to align with new ai.engine.models structure
- Updated import paths and root references in all affected files
