+++ gendox-core-api/src/main/java/dev/ctrlspace/gendox/gendoxcoreapi/ai/engine/model/dtos/generic/AiModelRequestParams.java
@@ -1,4 +1,4 @@
-package dev.ctrlspace.gendox.gendoxcoreapi.ai.engine.model.dtos;
+package dev.ctrlspace.gendox.gendoxcoreapi.ai.engine.model.dtos.generic;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
