+++ gendox-core-api/src/main/java/dev/ctrlspace/gendox/gendoxcoreapi/ai/engine/model/dtos/openai/request/OpenAiGptRequest.java
@@ -2,7 +2,7 @@ package dev.ctrlspace.gendox.gendoxcoreapi.ai.engine.model.dtos.openai.request;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
-import dev.ctrlspace.gendox.gendoxcoreapi.ai.engine.model.dtos.AiModelMessage;
+import dev.ctrlspace.gendox.gendoxcoreapi.ai.engine.model.dtos.generic.AiModelMessage;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
