+++ gendox-core-api/src/main/java/dev/ctrlspace/gendox/gendoxcoreapi/converters/AiModelMessagetoCohereGenerationsTextConverter.java
@@ -1,6 +1,6 @@
 package dev.ctrlspace.gendox.gendoxcoreapi.converters;
 
-import dev.ctrlspace.gendox.gendoxcoreapi.ai.engine.model.dtos.AiModelMessage;
+import dev.ctrlspace.gendox.gendoxcoreapi.ai.engine.model.dtos.generic.AiModelMessage;
 import dev.ctrlspace.gendox.gendoxcoreapi.ai.engine.model.dtos.cohere.response.CohereGenerations;
 
 public class AiModelMessagetoCohereGenerationsTextConverter {
